Improve project structure and extensibility
BREAKING: There is a new `-f` flag.  The old `-f` is now `-e`.

One of the big changes at this juncture is the separation of concerns and the addition of Dependency Injection.

I've done a lot of work to separate out concerns and make way for various ways I plan to expand on the base.

You'll notice:

* Scope Determiners - responsible for determining which files are in scope of processing
* A single DeclaredDependencyAggregator - which looks through those files and aggregates all Declared Dependencies
* The migration of how we interact with input and trigger the process into Controllers, to pave the way for non-composer plugin scripts
* The addition of several different Output Handlers for the default command (text, csv, json, and xml!)